CPM helper as separate component
